// web/src/features/routing/types.ts
/*
Copyright (C) 2023-2026 QuantumNous

This program is free software: you can redistribute it and/or modify
it under the terms of the GNU Affero General Public License as
published by the Free Software Foundation, either version 3 of the
License, or (at your option) any later version.

This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
GNU Affero General Public License for more details.

You should have received a copy of the GNU Affero General Public License
along with this program. If not, see <https://www.gnu.org/licenses/>.

For commercial licensing, please contact support@quantumnous.com
*/
export type RoutingSource = {
  id: string
  label: string
  description: string
  price_multiplier: number
  model_count: number
  is_default: boolean
}

export type RoutingFamily = {
  id: string
  label: string
  selected_source_id: string
  default_source_id: string
  sources: RoutingSource[]
  fallback_description: string
}

export type RoutingSourcesData = {
  families: RoutingFamily[]
  sources: RoutingSource[]
}

export type RoutingSourcesApiResponse = {
  success: boolean
  message?: string
  data?: RoutingSourcesData
}

export type RoutingSourceSelection = {
  familyId: string
  sourceId: string | null
}